+++ mayan/apps/document_parsing/settings.py
@@ -2,6 +2,8 @@ from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.smart_settings.classes import Namespace
 
+from .literals import DEFAULT_PDFTOTEXT_PATH
+
 namespace = Namespace(label=_('Document parsing'), name='document_parsing')
 
 setting_auto_parsing = namespace.add_setting(
@@ -12,7 +14,7 @@ setting_auto_parsing = namespace.add_setting(
 )
 setting_pdftotext_path = namespace.add_setting(
     global_name='DOCUMENT_PARSING_PDFTOTEXT_PATH',
-    default='/usr/bin/pdftotext',
+    default=DEFAULT_PDFTOTEXT_PATH,
     help_text=_(
         'File path to poppler\'s pdftotext program used to extract text '
         'from PDF files.'